Added composables and unit test for email
