--- car-showroom-11/src/components/User.vue
@@ -1,15 +1,0 @@
-<template>
-    <!-- <v-app>
-        <v-main>
-            <v-navigation-drawer permanent>
-                <h1 style="color: red;">Hii</h1>
-            </v-navigation-drawer>
-        </v-main>
-    </v-app> -->
-</template>
-
-<style>
-h1 {
-    font-size: 100px;
-}
-</style>
