+++ car-showroom-11/src/components/__tests__/pinia.spec.js
@@ -1,7 +1,7 @@
 import { setActivePinia, createPinia } from "pinia"
 import { test } from "vitest";
 import { useCarStore } from "../../stores/car"
-
+import axios from "axios";
 describe("userStore", () => {
     beforeEach(() => {
         setActivePinia(createPinia());
@@ -10,7 +10,10 @@ describe("userStore", () => {
         const carStore = useCarStore()
         expect(carStore.cars_info).toHaveLength(0)
         await carStore.getData()
-        expect(carStore.cars_info).toHaveLength(!0)
+        const res = await axios.get(`https://testapi.io/api/dartya/resource/cardata`)
+        const cars = carStore.cars_info = await res.data.data
+        const car_length = cars.length
+        expect(carStore.cars_info).toHaveLength(car_length)
 
     })
 })
