+++ car-showroom-11/src/composables/useFormattedDate.js
@@ -1,0 +1,8 @@
+export const useFormattedDate = () => {
+    const today = new Date();
+    const year = today.getFullYear();
+    const month = (today.getMonth() + 1).toString().padStart(2, '0'); // add 1 to adjust for 0-based indexing
+    const day = today.getDate().toString().padStart(2, '0');
+    const formattedDate = `${year}-${month}-${day}`;
+    return { formattedDate }
+}
